Add a prototype prep tool for marking up accented characters
This is a proof of concept implementation of a prep tool that scans
the text of a xml file for words with accented characters. It displays
all unique occurences and asks the user which ones are to be marked
up. The current implementation doesn't actually change the source file
and has a number of limitations:

* it is not re-entrant, i.e. it doesn't notice that a word has already
  been marked up.
* it doesn't show the context of the word